+++ Facebook.Unity/Mobile/MobileFacebookGameObject.cs
@@ -40,6 +40,11 @@ namespace Facebook.Unity.Mobile
             this.MobileFacebook.OnRefreshCurrentAccessTokenComplete(new ResultContainer(message));
         }
 
+        public void OnRefreshLimitedLoginComplete(string message)
+        {
+            this.MobileFacebook.OnRefreshLimitedLoginComplete(new ResultContainer(message));
+        }
+
         public void OnFriendFinderComplete(string message)
         {
             this.MobileFacebook.OnFriendFinderComplete(new ResultContainer(message));