+++ Facebook.Unity/PlatformEditor/EditorFacebook.cs
@@ -119,6 +119,16 @@ namespace Facebook.Unity.Editor
                 permissions.ToCommaSeparateList());
         }
 
+        public void RefreshLimitedLogin(
+            RefreshFallbackPolicy fallbackPolicy,
+            FacebookDelegate<ILoginResult> callback)
+        {
+            this.editorWrapper.ShowLoginMockDialog(
+                this.OnRefreshLimitedLoginComplete,
+                this.CallbackManager.AddFacebookDelegate(callback),
+                "public_profile");
+        }
+
         public override void AppRequest(
             string message,
             OGActionType? actionType,
@@ -377,6 +387,12 @@ namespace Facebook.Unity.Editor
         {
             var result = new AccessTokenRefreshResult(resultContainer);
             CallbackManager.OnFacebookResponse(result);
+        }
+
+        public void OnRefreshLimitedLoginComplete(ResultContainer resultContainer)
+        {
+            var result = new LoginResult(resultContainer);
+            this.OnAuthResponse(result);
         }
 
         public void OnFriendFinderComplete(ResultContainer resultContainer)